+++ php/Terminus/Models/Site.php
@@ -261,7 +261,7 @@ class Site extends TerminusModel {
   }
 
   /**
-   * Enables Redis caching
+   * Disables Redis caching
    *
    * @return array
    */
@@ -269,6 +269,20 @@ class Site extends TerminusModel {
     $response = $this->request->simpleRequest(
       'sites/' . $this->get('id') . '/settings',
       ['method' => 'put', 'form_params' => ['allow_cacheserver' => false]]
+    );
+    $this->convergeBindings();
+    return $response['data'];
+  }
+
+  /**
+   * Disables Solr indexing
+   *
+   * @return array
+   */
+  public function disableSolr() {
+    $response = $this->request->simpleRequest(
+      'sites/' . $this->get('id') . '/settings',
+      ['method' => 'put', 'form_params' => ['allow_indexserver' => false]]
     );
     $this->convergeBindings();
     return $response['data'];
@@ -283,6 +297,20 @@ class Site extends TerminusModel {
     $response = $this->request->simpleRequest(
       'sites/' . $this->get('id') . '/settings',
       ['method' => 'put', 'form_params' => ['allow_cacheserver' => true]]
+    );
+    $this->convergeBindings();
+    return $response['data'];
+  }
+
+  /**
+   * Enables Solr indexing
+   *
+   * @return array
+   */
+  public function enableSolr() {
+    $response = $this->request->simpleRequest(
+      'sites/' . $this->get('id') . '/settings',
+      ['method' => 'put', 'form_params' => ['allow_indexserver' => true]]
     );
     $this->convergeBindings();
     return $response['data'];